made a common game list component to use in the games screen and wishlist screen
