+++ screens/WishListScreen.tsx
@@ -1,73 +1,11 @@
 import { useContext } from 'react';
-import { View, StyleSheet, FlatList, Dimensions } from 'react-native';
-import * as WebBrowser from 'expo-web-browser';
 import { WishlistContext } from '../store/context/wishlist/wishlist-context';
-import { LinearGradient } from 'expo-linear-gradient';
-
-import { GameDealItem } from './GameDealsScreen';
-import GameDealCard from '../components/GameDealCard';
-import { CHEAPSHARK_REDIRECT_API } from '@env';
+import GameList from '../components/GameList';
 
 const WishListScreen = () => {
 	const wishListGames = useContext(WishlistContext);
-	const openBrowserAsync = async (dealID: string) => {
-		await WebBrowser.openBrowserAsync(`${CHEAPSHARK_REDIRECT_API}${dealID}`);
-	};
-	const renderCards = ({ item }: { item: GameDealItem }) => {
-		return (
-			<View style={styles.wishListItemContainer}>
-				<GameDealCard
-					gameDealItem={item}
-					handleGameDealPress={openBrowserAsync}
-				/>
-			</View>
-		);
-	};
 
-	return (
-		<View style={styles.linearGradient}>
-			<LinearGradient
-				style={styles.linearGradient}
-				colors={['#313131', '#dfdfdf', '#313131']}>
-				<View style={styles.listContainer}>
-					<>
-						<FlatList
-							data={wishListGames.games}
-							renderItem={renderCards}
-							contentContainerStyle={{ padding: 5 }}
-							numColumns={2}
-							keyExtractor={(item) => `${item.dealID}`}
-						/>
-					</>
-				</View>
-			</LinearGradient>
-		</View>
-	);
+	return <GameList games={wishListGames.games} />;
 };
 
 export default WishListScreen;
-const { width } = Dimensions.get('window');
-const styles = StyleSheet.create({
-	rootContainer: {
-		flex: 1,
-		alignItems: 'center',
-		justifyContent: 'center',
-		backgroundColor: '#282828',
-	},
-	linearGradient: {
-		flex: 1,
-	},
-	listContainer: {
-		flex: 1,
-		alignItems: 'center',
-		justifyContent: 'center',
-		elevation: 4,
-		shadowColor: 'black',
-		shadowOffset: { width: 4, height: 2 },
-		shadowRadius: 4,
-		shadowOpacity: 0.5,
-	},
-	wishListItemContainer: {
-		width: width / 2,
-	},
-});
